spi: tegra: Supporting HW based CS control
Supporting the hw based CS to communicate to spi device. This
provides the constraints in hold and setup time of CS before
clock start and clock ends.

The hw based CS can be selected if spi client provide the option
through the device controler data and only one transfer per
message is requested.

Reviewed-on: http://git-master/r/23988

Change-Id: Iba6d1efd8c5406f1b44ddd44b6f739ac529a94df
Reviewed-on: http://git-master/r/47150
Tested-by: Cho-Che Cheng <jacheng@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>